Renamed some folders and changed structure
