--- component.mk
@@ -1,4 +1,0 @@
-
-COMPONENT_ADD_INCLUDEDIRS := inc
-COMPONENT_PRIV_INCLUDEDIRS := inc/led_strip
-
