+++ examples/simple/Makefile
@@ -4,6 +4,5 @@
 #
 
 PROJECT_NAME := simple
-EXTRA_COMPONENT_DIRS := "../../../"
 
 include $(IDF_PATH)/make/project.mk
