+++ include/led_strip.h
